Moved bootstrap config functions into the utlis/files module
