+++ tests/fixtures/invalid_bootstrap_config.yml
