+++ tests/fixtures/valid_bootstrap_config.yml
